lesson 4.3.6, наследование
